test(sql): rename JavascriptContextTest to JavascriptContextBuilderTest
Mirror the class under test: move the test alongside JavascriptContextBuilder
in the org.molgenis.emx2.sql package, and drop the superseded single-test
processors/JavascriptContextBuilderTest stub, whose one test this file
already covers.
